Thu, Sep 7, 2023, 9:05 AM +01:00

+++ llamas.c
@@ -20,7 +20,7 @@ int main(void)
 
   int year = 0;
 
-  unrounded = startPo / 12;
+  unrounded = (float) startPo / 12;
 
   startPo = round(startPo / 12);
